Removes properties from git section
It seems that github cannot parse it properly. I used properties to give
custom id to heading so that I can create a table of contents.
Github does not parse it and it automatically creates a link for the heading
which is same as the headings title.

In this commit, I have used the github link.