Resumes tracking vundle directory
I added .vim/bundle to gitignore. I had to run `rm --cached` to remove
it. As a result vundle also got removed. As it did not appear in the
repo any longer, the clones did not have it in their repo.

It worked fine on my local copy because I had a copy of it. But my other
systems which cloned the repo did not get vundle with them.