Tmux: adds sleep to avoid launching two instances of MacVim
One instance of MacVim will have both the servers running: Vim and
Notes.
The other instance does not have any window or tab in it.

My guess is it is because mvim is launched in quick succession. I have
added a one second delay to fix this issue.

--nofork didn't help. But then the control does not return to the cursor
until that instance of MacVim is closed.