Tmux: Changes tmux default session
I don't need to open files in Vim. I can easily open them using ctrlP
and NERDTree.

I also want to change the panes in which ping and iftop runs. iftop
should run in the right pane and ping in the left pane. This way each
time I jump from top to bottom pane, I will drop in the ping pane. This
is the one that I usually have to restart unlike iftop.