Git: Fix a bug in pretty_git_log that will not display first commit in log
It happened b/c column command expects a newline at the end of the input
or text file. If it is missing it will give an error

```
column: line too long
```

I added a new line $FORMAT to make it work.

Also, sed and awk are poor alternative of column command as they do not
give columns with consistent width if text in a column has varying
length.