Vim: Remove configuration that close Vim if the only window left open is NERDTree
It closes window even if there are buffers present in the list. I tried
to tweak it with

```
echo len(filter(range(1,bufnr('$')), 'getbufvar(v:val, "&buflisted")'))
```

from Raimondi on #vim

But it's not worth it. I mean when will the use case occur that all
buffers have been closed and the only one remaining is NERDTree. It does
not happen, if you look for open buffers, b/c a single empty buffer always
stays open.